requirements.txt
# Django & Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8

# WebSocket & Real-time
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Database
psycopg2-binary==2.9.9  # PostgreSQL adapter

# Images & Vision Processing
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.26.2

# Fingerprint processing
scikit-image==1.28.0
scipy==1.11.3

# GPS & Geolocation
geopy==2.4.0
googlemaps==4.10.0

# Notifications
twilio==8.10.0
firebase-admin==6.2.0

# Email
django-anymail==10.1

# Celery for async tasks
celery==5.3.4
redis==5.0.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0

# Development
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
